+++ include/samplog/PluginLogger.hpp
@@ -43,7 +43,7 @@ namespace samplog
 	public:
 		virtual void Log(const LogLevel level, const char *msg) = 0;
 		virtual void Log(AMX * const amx, const LogLevel level, const char *msg) = 0;
-		virtual void LogEx(const LogLevel level, const char *msg, long line, const char *file, const char *function) = 0;
+		virtual void LogEx(const LogLevel level, const char *msg, int line, const char *file, const char *function) = 0;
 		virtual bool LogNativeCall(AMX * const amx, const char *name, const char *params_format) = 0;
 		virtual void SetLogLevel(const LogLevel level, bool enabled) = 0;
 		virtual bool IsLogLevel(const LogLevel log_level) = 0;
